fix : search box navigation
검색창 네비게이션 추가
