+++ 닉시_조승연/Week03_Mission/Week03_Mission/Cells/HomeCellView.swift
@@ -1,0 +1,49 @@
+import UIKit
+import SnapKit
+
+class HomeCellView: UICollectionViewCell {
+    
+    private let imageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFit
+        imageView.clipsToBounds = true
+        return imageView
+    }()
+    
+    private let titleLabel: UILabel = {
+        let label = UILabel()
+        label.font = UIFont.systemFont(ofSize: 12)
+        label.textColor = .black
+        label.textAlignment = .center
+        label.numberOfLines = 2 // 두 줄로 표시
+        return label
+    }()
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        contentView.addSubview(imageView)
+        contentView.addSubview(titleLabel)
+        
+        // 이미지 뷰 배치
+        imageView.snp.makeConstraints { make in
+            make.top.equalToSuperview()
+            make.centerX.equalToSuperview()
+            make.width.height.equalTo(60)
+        }
+        
+        // 타이틀 레이블 배치
+        titleLabel.snp.makeConstraints { make in
+            make.top.equalTo(imageView.snp.bottom).offset(5)
+            make.leading.trailing.equalToSuperview()
+        }
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func configure(with image: UIImage, title: String) {
+        imageView.image = image
+        titleLabel.text = title
+    }
+}
