+++ 닉시_조승연/Week03_Mission/Week03_Mission/VIewControllers/SimpleSearchViewController.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+class SimpleSearchViewController: UIViewController {
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        view.backgroundColor = .white
+        title = "검색" 
+    }
+}
